+++ content/table.mjs.js
@@ -162,7 +162,7 @@ export default class DataTable {
 	}
 
 	/**
-	 * @param {Object<string, string|number|boolean>[]} data
+	 * @param {{[x: string]: string|number|boolean}[]} data
 	 * @param {string[]} [sortOrder]
 	 * @returns {void}
 	 */