+++ modules/authVerifier.mjs.js
@@ -18,15 +18,15 @@
 
 export const moduleVersion = "2.0.0";
 
-import { addrIsInDomain2, domainIsInDomain, getDomainFromAddr } from "./utils.mjs.js";
+import { addrIsInDomain, addrIsInDomain2, domainIsInDomain, getDomainFromAddr } from "./utils.mjs.js";
 import ArhParser from "./arhParser.mjs.js";
-import { DKIM_InternalError } from "./error.mjs.js";
 import DMARC from "./dkim/dmarc.mjs.js";
 import ExtensionUtils from "./extensionUtils.mjs.js";
 import Logging from "./logging.mjs.js";
 import MsgParser from "./msgParser.mjs.js";
 import SignRules from "./dkim/signRules.mjs.js";
 import Verifier from "./dkim/verifier.mjs.js";
+import { getBimiIndicator } from "./bimi.mjs.js";
 import { getFavicon } from "./dkim/favicon.mjs.js";
 import prefs from "./preferences.mjs.js";
 
@@ -51,11 +51,12 @@ const log = Logging.getLogger("AuthVerifier");
 
 /**
  * @typedef {object} SavedAuthResultV3
- * @property {string} version Result version ("3.0").
+ * @property {string} version Result version ("3.1").
  * @property {dkimSigResultV2[]} dkim
  * @property {ArhResInfo[]|undefined} [spf]
  * @property {ArhResInfo[]|undefined} [dmarc]
  * @property {{dkim?: dkimSigResultV2[]}|undefined} [arh]
+ * @property {string|undefined} [bimiIndicator] Since version 3.1
  */
 /**
  * @typedef {SavedAuthResultV3} SavedAuthResult
@@ -112,12 +113,32 @@ export default class AuthVerifier {
 		// check for saved AuthResult
 		let savedAuthResult = await loadAuthResult(message);
 		if (savedAuthResult) {
-			return SavedAuthResult_to_AuthResult(savedAuthResult);
+			let from = null;
+			try {
+				from = MsgParser.parseAuthor(message.author, prefs["internationalized.enable"]);
+			} catch (error) {
+				log.warn("Parsing of from header failed", error);
+			}
+			return SavedAuthResult_to_AuthResult(savedAuthResult, from);
 		}
 
 		// create msg object
 		const rawMessage = await browser.messages.getRaw(message.id);
-		const msgParsed = MsgParser.parseMsg(rawMessage);
+		let msgParsed;
+		try {
+			msgParsed = MsgParser.parseMsg(rawMessage);
+		} catch (error) {
+			log.error("Parsing of message failed", error);
+			return Promise.resolve({
+				version: "2.1",
+				dkim: [{
+					version: "2.0",
+					result: "PERMFAIL",
+					res_num: 30,
+					result_str: browser.i18n.getMessage("DKIM_INTERNALERROR_INCORRECT_EMAIL_FORMAT"),
+				}],
+			});
+		}
 		const fromHeader = msgParsed.headers.get("from");
 		if (!fromHeader || !fromHeader[0]) {
 			throw new Error("message does not contain a from header");
@@ -160,13 +181,14 @@ export default class AuthVerifier {
 				savedAuthResult = arhResult;
 			} else {
 				savedAuthResult = {
-					version: "3.0",
+					version: "3.1",
 					dkim: [],
 					spf: arhResult.spf,
 					dmarc: arhResult.dmarc,
 					arh: {
 						dkim: arhResult.dkim
 					},
+					bimiIndicator: arhResult.bimiIndicator,
 				};
 			}
 		} else {
@@ -194,7 +216,7 @@ export default class AuthVerifier {
 		saveAuthResult(message, savedAuthResult).
 			catch(error => log.fatal("Failed to store result", error));
 
-		const authResult = await SavedAuthResult_to_AuthResult(savedAuthResult);
+		const authResult = await SavedAuthResult_to_AuthResult(savedAuthResult, msg.from);
 		log.debug("authResult: ", authResult);
 		return authResult;
 	}
@@ -234,6 +256,8 @@ async function getARHResult(message, headers, from, listId, account, dmarc) {
 	let arhSPF = [];
 	/** @type {ArhResInfo[]} */
 	let arhDMARC = [];
+	/** @type {ArhResInfo[]} */
+	let arhBIMI = [];
 	for (const header of arHeaders) {
 		/** @type {import("./arhParser.mjs.js").ArhHeader} */
 		let arh;
@@ -270,6 +294,9 @@ async function getARHResult(message, headers, from, listId, account, dmarc) {
 		arhDMARC = arhDMARC.concat(arh.resinfo.filter((element) => {
 			return element.method === "dmarc";
 		}));
+		arhBIMI = arhBIMI.concat(arh.resinfo.filter((element) => {
+			return element.method === "bimi";
+		}));
 	}
 
 	// convert DKIM results
@@ -302,7 +329,7 @@ async function getARHResult(message, headers, from, listId, account, dmarc) {
 					case 2: // ignore
 						break;
 					default:
-						throw new DKIM_InternalError("invalid error.algorithm.sign.rsa-sha1.treatAs");
+						throw new Error("invalid error.algorithm.sign.rsa-sha1.treatAs");
 				}
 			}
 		}
@@ -312,10 +339,11 @@ async function getARHResult(message, headers, from, listId, account, dmarc) {
 	sortSignatures(dkimSigResults, from, listId);
 
 	const savedAuthResult = {
-		version: "3.0",
+		version: "3.1",
 		dkim: dkimSigResults,
 		spf: arhSPF,
 		dmarc: arhDMARC,
+		bimiIndicator: getBimiIndicator(headers, arhBIMI) ?? undefined,
 	};
 	log.debug("ARH result:", savedAuthResult);
 	return savedAuthResult;
@@ -338,8 +366,8 @@ async function saveAuthResult(message, savedAuthResult) {
 		// reset result
 		log.debug("reset AuthResult result");
 		await browser.storageMessage.set(message.id, "dkim_verifier@pl-result", "");
-	} else if (savedAuthResult.dkim[0]?.result === "TEMPFAIL") {
-		// don't save result if DKIM result is a TEMPFAIL
+	} else if (savedAuthResult.dkim.some(res => res.result === "TEMPFAIL")) {
+		// don't save result if DKIM result contains a TEMPFAIL
 		log.debug("result not saved because DKIM result is a TEMPFAIL");
 	} else {
 		log.debug("save AuthResult result");
@@ -487,7 +515,7 @@ function sortSignatures(signatures, from, listId) {
 			return 1;
 		}
 
-		throw new DKIM_InternalError(`result_compare: sig1.result: ${sig1.result}; sig2.result: ${sig2.result}`);
+		throw new Error(`result_compare: sig1.result: ${sig1.result}; sig2.result: ${sig2.result}`);
 	}
 
 	/**
@@ -638,7 +666,7 @@ function arhDKIM_to_dkimSigResultV2(arhDKIM) {
 			}
 			break;
 		default:
-			throw new DKIM_InternalError(`invalid dkim result in arh: ${arhDKIM.result}`);
+			throw new Error(`invalid dkim result in arh: ${arhDKIM.result}`);
 	}
 	let sdid = arhDKIM.propertys.header.d;
 	let auid = arhDKIM.propertys.header.i;
@@ -693,7 +721,6 @@ function dkimResultV1_to_dkimSigResultV2(dkimResultV1) {
  *
  * @param {dkimSigResultV2} dkimSigResult
  * @returns {AuthResultDKIM}
- * @throws DKIM_InternalError
  */
 function dkimSigResultV2_to_AuthResultDKIM(dkimSigResult) { // eslint-disable-line complexity
 	/** @type {AuthResultDKIM} */
@@ -826,7 +853,7 @@ function dkimSigResultV2_to_AuthResultDKIM(dkimSigResult) { // eslint-disable-li
 			authResultDKIM.result_str = browser.i18n.getMessage("NOSIG");
 			break;
 		default:
-			throw new DKIM_InternalError(`unknown result: ${dkimSigResult.result}`);
+			throw new Error(`unknown result: ${dkimSigResult.result}`);
 	}
 
 	return authResultDKIM;
@@ -836,9 +863,10 @@ function dkimSigResultV2_to_AuthResultDKIM(dkimSigResult) { // eslint-disable-li
  * Convert SavedAuthResult to AuthResult.
  *
  * @param {SavedAuthResult} savedAuthResult
+ * @param {string?} from
  * @returns {Promise<AuthResult>} authResult
  */
-function SavedAuthResult_to_AuthResult(savedAuthResult) {
+function SavedAuthResult_to_AuthResult(savedAuthResult, from) {
 	/** @type {AuthResult} */
 	const authResult = {
 		version: "2.1",
@@ -856,7 +884,7 @@ function SavedAuthResult_to_AuthResult(savedAuthResult) {
 				dkimSigResultV2_to_AuthResultDKIM)
 		};
 	}
-	return addFavicons(authResult);
+	return addFavicons(authResult, from, savedAuthResult.bimiIndicator);
 }
 
 /**
@@ -884,9 +912,11 @@ function AuthResultDKIMV2_to_dkimSigResultV2(authResultDKIM) {
  * Add favicons to the DKIM results.
  *
  * @param {AuthResult} authResult
+ * @param {string?} from
+ * @param {string|undefined} bimiIndicator
  * @returns {Promise<AuthResult>} authResult
  */
-async function addFavicons(authResult) {
+async function addFavicons(authResult, from, bimiIndicator) {
 	if (!prefs["display.favicon.show"]) {
 		return authResult;
 	}
@@ -895,7 +925,11 @@ async function addFavicons(authResult) {
 	}
 	for (const dkim of authResult.dkim) {
 		if (dkim.sdid) {
-			dkim.favicon = await getFavicon(dkim.sdid);
+			if (bimiIndicator && from && addrIsInDomain(from, dkim.sdid)) {
+				dkim.favicon = `data:image/svg+xml;base64,${bimiIndicator}`;
+			} else {
+				dkim.favicon = await getFavicon(dkim.sdid, dkim.auid, from);
+			}
 		}
 	}
 	return authResult;