+++ modules/dkim/favicon.mjs.js
@@ -10,30 +10,46 @@
 // @ts-check
 /* eslint-env webextensions */
 
+import { addrIsInDomain, domainIsInDomain } from "../utils.mjs.js";
 import ExtensionUtils from "../extensionUtils.mjs.js";
-import { domainIsInDomain } from "../utils.mjs.js";
 
-/** @type {Object<string, string|undefined>} */
+/** @type {{[x: string]: string|undefined}} */
 let favicons;
 
 /**
  * Get the URL to the favicon of the sdid, if available.
  *
  * @param {string} sdid
+ * @param {string|undefined} auid
+ * @param {string?} from
  * @returns {Promise<string|undefined>} url to favicon
  */
-export async function getFavicon(sdid) {
+export async function getFavicon(sdid, auid, from) {
 	if (!favicons) {
 		const faviconsStr = await ExtensionUtils.readFile("data/favicon.json");
 		// eslint-disable-next-line require-atomic-updates
 		favicons = JSON.parse(faviconsStr);
 	}
 
-	let url = favicons[sdid];
+	// Check if enabled for SDID.
+	let url = favicons[sdid.toLowerCase()];
 	if (!url) {
+		// Check if enabled for the base domain of the SDID.
 		const baseDomain = Object.keys(favicons).find(domain => domainIsInDomain(sdid, domain));
 		if (baseDomain) {
 			url = favicons[baseDomain];
+		}
+	}
+	if (!url) {
+		// Check if enabled for AUID.
+		if (auid) {
+			url = favicons[auid.toLowerCase()];
+		}
+	}
+	if (!url) {
+		// Check if enabled for from.
+		if (from && addrIsInDomain(from, sdid)) {
+			url = favicons[from.toLowerCase()];
 		}
 	}